fix(harness): keep map hashing server-side
Refs: SAP-3149
